+++ CHANGELOG.md
@@ -4,14 +4,19 @@
 
 ## [Unreleased]
 
+## [0.4.0]
+### Added
+- [#8](https://github.com/picimako/mockitools/issues/8): Added a new inspection that reports multiple consecutive calls on `*Throw()` stubbing calls.
+They can be merged into a single such call.
+
 ## [0.3.0]
 - **IntelliJ version support: versions prior to 2021.1 are no longer supported.**
 
 ### Added
 - Added a new inspection to report checked exceptions in `*Throw()` stubbing calls that are not specified in the stubbed method's `throws` clause.
 See [documentation](/docs/stubbing.md#invalid-checked-exception-is-passed-into-throw-methods).
-- [#8](https://github.com/picimako/mockitools/issues/8): Added a new inspection that reports multiple consecutive calls on `*Return()` stubbing calls. These can be merged into a single such call.
-See [documentation](docs/stubbing.md#consecutive-return-calls-can-be-merged).
+- [#8](https://github.com/picimako/mockitools/issues/8): Added a new inspection that reports multiple consecutive calls on `*Return()` stubbing calls.
+These can be merged into a single such call. See [documentation](docs/stubbing.md#consecutive-return-and-throw-calls-can-be-merged).
 
 ### Changed
 - Updated Gradle IntelliJ plugin to 1.4.0, gradle to 7.4, and qodana-action to 4.2.5.